+++ internal/app/grpc/bufdownload/download.go
@@ -6,8 +6,8 @@ import (
 	"buf.build/gen/go/bufbuild/registry/connectrpc/go/buf/registry/module/v1/modulev1connect"
 	modulev1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/module/v1"
 	"connectrpc.com/connect"
+	"github.com/alipourhabibi/Hades/internal/pkg/services/module"
 	"github.com/alipourhabibi/Hades/models"
-	"github.com/alipourhabibi/Hades/internal/pkg/services/bufdownload"
 	"github.com/alipourhabibi/Hades/utils/log"
 	"github.com/google/uuid"
 )
@@ -15,11 +15,11 @@ import (
 type Server struct {
 	modulev1connect.DownloadServiceHandler
 
-	service *bufdownload.Service
+	service *module.Service
 	logger  *log.LoggerWrapper
 }
 
-func NewServer(l *log.LoggerWrapper, service *bufdownload.Service) *Server {
+func NewServer(l *log.LoggerWrapper, service *module.Service) *Server {
 	return &Server{
 		logger:  l,
 		service: service,
