+++ internal/app/grpc/bufgraph/graph.go
@@ -7,8 +7,8 @@ import (
 	"buf.build/gen/go/bufbuild/registry/connectrpc/go/buf/registry/module/v1/modulev1connect"
 	modulev1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/module/v1"
 	"connectrpc.com/connect"
+	"github.com/alipourhabibi/Hades/internal/pkg/services/module"
 	"github.com/alipourhabibi/Hades/models"
-	"github.com/alipourhabibi/Hades/internal/pkg/services/bufgraph"
 	"github.com/alipourhabibi/Hades/utils/log"
 	"github.com/google/uuid"
 )
@@ -16,11 +16,11 @@ import (
 type Server struct {
 	modulev1connect.GraphServiceHandler
 
-	service *bufgraph.Service
+	service *module.Service
 	logger  *log.LoggerWrapper
 }
 
-func NewServer(l *log.LoggerWrapper, service *bufgraph.Service) *Server {
+func NewServer(l *log.LoggerWrapper, service *module.Service) *Server {
 	return &Server{
 		logger:  l,
 		service: service,
