+++ internal/app/grpc/bufmodules/bufmodules.go
@@ -6,19 +6,19 @@ import (
 	moduleConnV1 "buf.build/gen/go/bufbuild/registry/connectrpc/go/buf/registry/module/v1/modulev1connect"
 	modulev1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/module/v1"
 	"connectrpc.com/connect"
+	"github.com/alipourhabibi/Hades/internal/pkg/services/module"
 	"github.com/alipourhabibi/Hades/models"
-	"github.com/alipourhabibi/Hades/internal/pkg/services/bufmodules"
 	"github.com/alipourhabibi/Hades/utils/log"
 )
 
 type Server struct {
 	moduleConnV1.ModuleServiceHandler
 
-	service *bufmodules.Service
+	service *module.Service
 	logger  *log.LoggerWrapper
 }
 
-func NewServer(l *log.LoggerWrapper, service *bufmodules.Service) *Server {
+func NewServer(l *log.LoggerWrapper, service *module.Service) *Server {
 	return &Server{
 		logger:  l,
 		service: service,
