+++ internal/pkg/services/authentication/authentication.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"errors"
 
-	"github.com/alipourhabibi/Hades/models"
 	pkgerr "github.com/alipourhabibi/Hades/internal/pkg/errors"
 	"github.com/alipourhabibi/Hades/internal/pkg/services/authorization"
 	"github.com/alipourhabibi/Hades/internal/storage/db/session"
 	sessiondb "github.com/alipourhabibi/Hades/internal/storage/db/session"
 	userdb "github.com/alipourhabibi/Hades/internal/storage/db/user"
+	"github.com/alipourhabibi/Hades/models"
 	"github.com/alipourhabibi/Hades/utils/bcrypt"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
